Makes vorgan/identity/etc stuff not transfer over to an occupied tf_form

+++ modular_chomp/code/modules/mob/living/living.dm
@@ -124,36 +124,37 @@ Maybe later, gotta figure out a way to click yourself when in a locker etc.
 		new_mob.faction = src.faction
 
 		if(new_mob && isliving(new_mob))
-			for(var/obj/belly/B as anything in new_mob.vore_organs)
-				new_mob.vore_organs -= B
-				qdel(B)
-			new_mob.vore_organs = list()
-			new_mob.name = src.name
-			new_mob.real_name = src.real_name
-			for(var/lang in src.languages)
-				new_mob.languages |= lang
-			src.copy_vore_prefs_to_mob(new_mob)
-			new_mob.vore_selected = src.vore_selected
-			if(ishuman(src))
-				var/mob/living/carbon/human/H = src
-				if(ishuman(new_mob))
-					var/mob/living/carbon/human/N = new_mob
-					N.gender = H.gender
-					N.identifying_gender = H.identifying_gender
+			if(!new_mob.ckey)
+				for(var/obj/belly/B as anything in new_mob.vore_organs)
+					new_mob.vore_organs -= B
+					qdel(B)
+				new_mob.vore_organs = list()
+				new_mob.name = src.name
+				new_mob.real_name = src.real_name
+				for(var/lang in src.languages)
+					new_mob.languages |= lang
+				src.copy_vore_prefs_to_mob(new_mob)
+				new_mob.vore_selected = src.vore_selected
+				if(ishuman(src))
+					var/mob/living/carbon/human/H = src
+					if(ishuman(new_mob))
+						var/mob/living/carbon/human/N = new_mob
+						N.gender = H.gender
+						N.identifying_gender = H.identifying_gender
+					else
+						new_mob.gender = H.gender
 				else
-					new_mob.gender = H.gender
-			else
-				new_mob.gender = src.gender
-				if(ishuman(new_mob))
-					var/mob/living/carbon/human/N = new_mob
-					N.identifying_gender = src.gender
+					new_mob.gender = src.gender
+					if(ishuman(new_mob))
+						var/mob/living/carbon/human/N = new_mob
+						N.identifying_gender = src.gender
 
-			for(var/obj/belly/B as anything in src.vore_organs)
-				B.loc = new_mob
-				B.forceMove(new_mob)
-				B.owner = new_mob
-				src.vore_organs -= B
-				new_mob.vore_organs += B
+				for(var/obj/belly/B as anything in src.vore_organs)
+					B.loc = new_mob
+					B.forceMove(new_mob)
+					B.owner = new_mob
+					src.vore_organs -= B
+					new_mob.vore_organs += B
 
 			new_mob.ckey = src.ckey
 			if(new_mob_ckey)
